Finished Icon infos, begin player slot box work
Icon info windows are finished for the time being, but still need to get
usage gifs added as well as necessary information. The current decision
is to add an "A" to the icons of weapons on S, and different gifs
showing what's different. An "A" will likely be overlayed on the gif as
well.New player box added in, all imagework done for player slots, and
"open" status done.