Player Slot Changes
Player slots now have effects to show when a player joins, leaves/is
kicked, readies up/is ready, and when the slot is opened or closed. Each
slot is managed by an icon next to it. Team change is handled by one of
these now.
I still want the slots to reorganize themselves depending on who is on
what team.