Waiting Room 2
Player slot work progression, I still want to change the way they look
exactly, though, and perhaps have a mouseover tip for what happens when
you click them.
The map box has been placed, but the internals need to be moved, and the
map preview backgrounds need to be made.
The font still needs to be updated.