Waiting Room 4
Map Preview box finished.
Modification box finished.
Text is beginning to be converted to a better font.
Player slots need to be updated still.
A few graphical errors need to be fixed still, such as the playerslot
glow appearing when exiting the room.
The entire room needs to be fixed for the clients too.