obj_mod code cleanup 2
Client and Host mod objects have the end steps updated.
The begin step has started to be updated, with the refresh packet
netcode completed.
There's a little bit of sidework to be done:
Siphon uses current_target_object
bullet_create needs to work regardless of the source(not just
obj_player)
primary_direction handled in the mouse press

Other than this, ship spawns and deaths need to still be handled