Debugging 2
More menu issues have been fixed- chat, etc.
Ships now spawn and can be moved, but netcode doesn't seem to be keeping
up with the movement.
Primaries can be fired, but collisions aren't happening correctly.
When secondaries are fired, the dummy is sending communications and
glitching out due to the mid being wrong.